Relocates placeholder modules
Moves placeholder modules from `core/grpc` and `core/interfaces` to a dedicated `core/placeholders` directory.

This change improves project structure and clarifies the purpose of these modules.
